+++ src/Cucumber.Pro.SpecFlowPlugin/Publishing/CucumberProResultsUrlBuilder.cs
@@ -13,13 +13,28 @@ namespace Cucumber.Pro.SpecFlowPlugin.Publishing
         public static string BuildCucumberProUrl(Config config, string projectName)
         {
             var cucumberProUrl = GetCucumberProUrl(config);
-            return $"{cucumberProUrl}tests/results/{projectName}";
+            return BuildCucumberProUrl(cucumberProUrl, projectName);
         }
 
-        private static String GetCucumberProUrl(Config config)
+        public static string BuildCucumberProUrl(string cucumberProUrl, string projectName)
+        {
+            return $"{cucumberProUrl}tests/results/{EncodeUriComponent(projectName)}";
+        }
+
+        private static string GetCucumberProUrl(Config config)
         {
             var cucumberProUrl = config.GetString(ConfigKeys.CUCUMBERPRO_URL);
             return !cucumberProUrl.EndsWith("/") ? cucumberProUrl + "/" : cucumberProUrl;
+        }
+
+        private static string EncodeUriComponent(string s) 
+        {
+            return Uri.EscapeDataString(s)
+                            .Replace("%21", "!")
+                            .Replace("%27", "'")
+                            .Replace("%28", "(")
+                            .Replace("%29", ")");
+
         }
     }
 }